requirements.txt
gallium>=1.5.1
imagination>=3.1.0
kotoba
cryptography
keyring
pyperclip
xmode>=0.5
grpcio-tools
grpcio
qrcode[pil]
pyjwt
sqlalchemy
pymysql